+++ persistent/ChangeLog.md
@@ -2,13 +2,15 @@
 
 ## 2.12.1.1
 
+* [#1231](https://github.com/yesodweb/persistent/pull/1231)
+    * Simplify Line type in Quasi module, always use NonEmpty
 * [#1229](https://github.com/yesodweb/persistent/pull/1229)
     * The `#id` labels are now generated for entities.
 
 ## 2.12.1.0
 
 * [#1218](https://github.com/yesodweb/persistent/pull/1218)
-    * Refactoring name generating functions in TH 
+    * Refactoring name generating functions in TH
 * [#1226](https://github.com/yesodweb/persistent/pull/1226)
     * Expose the `filterClause` and `filterClauseWithValues` functions to support
       the `upsertWhere` functionality in `persistent-postgresql`.